Move age::FileKey to age_core::format::FileKey
