+++ age/src/keys.rs
@@ -1,6 +1,6 @@
 //! Key structs and serialization.
 
-use age_core::primitives::hkdf;
+use age_core::{format::FileKey, primitives::hkdf};
 use rand::{rngs::OsRng, RngCore};
 use secrecy::{ExposeSecret, Secret};
 
@@ -14,47 +14,30 @@ use crate::{
 const HEADER_KEY_LABEL: &[u8] = b"header";
 const PAYLOAD_KEY_LABEL: &[u8] = b"payload";
 
-/// A file key for encrypting or decrypting an age file.
-pub struct FileKey(Secret<[u8; 16]>);
-
-impl From<[u8; 16]> for FileKey {
-    fn from(file_key: [u8; 16]) -> Self {
-        FileKey(Secret::new(file_key))
-    }
+pub(crate) fn new_file_key() -> FileKey {
+    let mut file_key = [0; 16];
+    OsRng.fill_bytes(&mut file_key);
+    file_key.into()
 }
 
-impl ExposeSecret<[u8; 16]> for FileKey {
-    fn expose_secret(&self) -> &[u8; 16] {
-        self.0.expose_secret()
-    }
+pub(crate) fn mac_key(file_key: &FileKey) -> HmacKey {
+    HmacKey(Secret::new(hkdf(
+        &[],
+        HEADER_KEY_LABEL,
+        file_key.expose_secret(),
+    )))
 }
 
-impl FileKey {
-    pub(crate) fn generate() -> Self {
-        let mut file_key = [0; 16];
-        OsRng.fill_bytes(&mut file_key);
-        file_key.into()
-    }
+pub(crate) fn v1_payload_key(
+    file_key: &FileKey,
+    header: &HeaderV1,
+    nonce: &Nonce,
+) -> Result<PayloadKey, Error> {
+    // Verify the MAC
+    header.verify_mac(mac_key(file_key))?;
 
-    pub(crate) fn mac_key(&self) -> HmacKey {
-        HmacKey(Secret::new(hkdf(
-            &[],
-            HEADER_KEY_LABEL,
-            self.0.expose_secret(),
-        )))
-    }
-
-    pub(crate) fn v1_payload_key(
-        &self,
-        header: &HeaderV1,
-        nonce: &Nonce,
-    ) -> Result<PayloadKey, Error> {
-        // Verify the MAC
-        header.verify_mac(self.mac_key())?;
-
-        // Return the payload key
-        Ok(PayloadKey(
-            hkdf(nonce.as_ref(), PAYLOAD_KEY_LABEL, self.0.expose_secret()).into(),
-        ))
-    }
+    // Return the payload key
+    Ok(PayloadKey(
+        hkdf(nonce.as_ref(), PAYLOAD_KEY_LABEL, file_key.expose_secret()).into(),
+    ))
 }
